fix: Add workaround for mkdocstring and factory-boy issue
